Update dataset_utils.py with various features
Added custom spectrogram resolution global parameter and updated the
appropriate methods for loading in the new data. Reworked the paths that
spectrogram images get saved to in a way that pytorch can easily
recognise and read in the data. Updated various helper methods and
linted where convenient. Also added temporary funtionality for limiting
the dataset to a binary choice for testing and experimentation purposes.